EOS-10334: Provide a way to mkfs Motr MD store (#66)
In case of Motr meta data corruption there's a requirement
to mkfs the existing Motr meta data store as part of the
recovery process of LDR-1 cluster.

Solution:
Extend prov-m0-reset script to support mkfs option inorder to format
motr meta data store. The script with mkfs option stops runs in offline
mode and does not require cluster to be running. The script temporaily
starts the required Consul and hax services to support motr mkfs and
does not affect the exiting cluster configuration.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>