EOS-12112: hax service stop timeout is less than the systemd timeout
As hax pacemaker resource stop timeout is less than the default systemd
timeout of 90 secs, hax service does not stop on time in case of situations
where it is unable to unmount /var/motr. This is typically due to ext4 data
sync operation in progress. This situation is observed in case of losing SAS
connection. In normal cases this timeout suffices. Unable to stop hax service
in time leads to stonith action.
Note that we purposefully do not set systemd stop timeout to allow clean
unmount of /var/motr and also because this issue is not seen in regular
failover/failback.

Solution:
Increase hax resource timeout to be more than systemd default timeout.