Problem: build-ha-io restore logic is broken (#113)
Restore function was added to support node replacement backup and restore. As
part of this there was a bug introduced due to missing and condition for
bootstrap calls. This was leading to invocation of restore function during
bootstrap.

Solution:
Add an and condition while checking for restore_conf flag in bootstrap path
as well as update path.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>