build-ha-io: support different interfaces on the nodes
Currently, the build-ha-io script expects that the name of the
data network interface is the same on both nodes, which sometimes
may not be the case when the underlying HW is different. (Which
may happen, for example, when the old failed node is replaced
with the new one.)

Solution: add support for different interface names via the
two new configuration options: `--left-iface` and `--right-iface`.
If the options are not explicitly set by user, the value of
the old `--iface` option is used for both interfaces.

Closes EOS-13356.

Reviewed-by: Ajay Srivastava <ajay.srivastava@seagate.com>
Reviewed-by: Mandar Sawant <mandar.sawant@seagate.com>
Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>