EOS-13529: components.ha.iostack-ha.refresh_config fails (#135)
As part of the node restore operation, backup of peer node was also being
created during restore. This was triggering restore from replaced node
to the active node as well. As during restore operation consul db is
restored, this leads to consul crash if consul is active on the node
leading to other service failures.

Solution:
- During restore peform backup only after completing the restore operation.
- Simplify node_backup() and node_restore()
- Create /var/lib/hare/ on replaced node if does not exists already during
  restore.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>