EOS-12635: After unmaintenance make sure consul is active before the export (#138)
As part of cortx software update process and during post update, the
cluster maintenance mode is disabled and provisioner immediately invokes
the build-ees-ha-update script. The script as a first step exports
consul kv which is also a most important step. Now if the Consul is not
active yet the consul export may fail creating a wrong exported file.
This leads to further issues as the key values go missing during import
leading to services failure.

Solution:
- Wait until Consul is active after the maintenance mode is disabled and before
  exporting Consul kv.
- Commit pacemaker CIB after io-stack is updated.
- Import exported consul kv before sspl update for sspl service to start cleanly.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>