Problem: prov-m0-reset --mkfs-only does not provide option to run on both nodes
Due to lvm used for Motr meta data segments running mkfs on the volumes on
single node does not reflect the changes on peer node when data recovery is
executed on both the nodes in parallel. Presently prov-m0-reset --mkfs-only
execution does not provide an option to run the operation on both the nodes.

Solution:
- Provide and option for the user to run m0mkfs for both the nodes on single node
  or on individual nodes.
- Incorporate network interface variable changes as per the latest main.
- Incorporate /var/motr directory related changes as per the latest main.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>