Problem: prepare_var_motr_dirs() may fail thus exiting the script
prepare_var_motr_dirs() checks on both the nodes if /var/motr1 and
/var/motr2 dirs are present, if not then renames the /var/motr to
/var/motr1 or /var/motr2 on respecive nodes.
While executing this command on the peer node the check condition
may fail if directories already exist thus leads to failure of the
ssh command for the peer node. This leads to script exit as `pipefail`
is set for the script.

Solution:
Make the directory creation command no-op by returning true in case of
check condition failure to continue the build script.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>