EOS-13296: haproxy service does not restart automatically on failure
If a pacemaker resource fails, pacemaker will try to restart the resource but if the
resource fails repeatedly then the failcount set by the pacemaker for the resource
needs to be cleared using `pcs resource cleanup`.
Setting failure-timeout meta attribute allows failcounts to expire automatically so
pacemaker tries to restart the resource.

Solution:
Set failure-timeout meta attribute of the haproxy pacemaker resource inorder to allow
automatic expiration of haproxy failcounts to trigger restart.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>