Problem: ClusterIP resource may sporadically fail and migrate to another node (#196)
Since the reason is unknown so far, this patch is supposed to mitigate
the problem making pacemaker to retry ClusterIP resource on banned node
once in 5 minutes.
Nothing depends from ClusterIP resource, so it is not supposed to cause
restart of other resources.
Timeout value is chosen in order to not disturb cluster too frequently.

The consequences of suggested patch are:
* Once in 5 minutes ClusterIP failures will be cleaned from 'pcs status'
  log. However they persist in syslog and corosync log.
  In case resource is not recovered successfully 'pcs status' will
  continue to show failed status soon.

Signed-off-by: Andrei Zheregelia <andrei.zheregelia@seagate.com>
(cherry picked from commit f267fd788cefad42e54a90c47724fc1ce4897a3a)