Fix mandatory arguments handling in build-ha-update
`build-ha-update` script expects three configuration files to be
provided as command line arguments but it didn't check whether the files
existed or not.

That could lead to the use of uninitialized variables. For example
'lnode' and 'rnode' variables, that are used in 'wait4consul()' might've
been unbound, which resulted in the infinite 'while' loop.