EOS-19478: RPM package for cortx-validator
Approach is the same as for main cortx-ha package with exception that
only one python package is targeted.

build.sh builds this package if "-t" option is passed.
build_validator.sh is kept in case standalone script is required, though
probably better to build cortx-validator package only in build.sh

Signed-off-by: Andrei Zheregelia <andrei.zheregelia@seagate.com>