CORTX-27961: [ConfStore search API] Use SERVICE_S3_SERVER instead of SERVICE_S3_HAPROXY to identify server pods
HA is not able to detect events for server nodes because monitor service
requires list of ids(machine-ids) that can be monitored. List does not contain
server node machine-ids because there is a change in cluster.conf. Previously,
cluster.conf has ha-proxy service listed which was used to identify server pods.
But, now, s3 is replaced by rgw now and there will not be any services in rgw as
of now.

Solution: Use COMPONENT_RGW constant defined in cortx-utils to identify server pod.
server pod config will be like this:
 9f4033f2f0304859b175ae9680c28a8d:
    cluster_id: 1a3acac8233d425c8dbfad522a99feca
    components:
    - name: utils
    - name: hare
    - name: rgw
    hostname: cortx-server-headless-svc-ssc-vm-g4-rhev4-0582
    name: cortx-server-headless-svc-ssc-vm-g4-rhev4-0582
    num_components: 3
    storage_set: storage-set-1
    type: server_node
so, instead of "services", "name" will be provided in recursive search pattern of ConfStoreSearch

Signed-off-by: Madhura Mande <madhura.mande@seagate.com>