CORTX-29933: System health update for control node (#703)
* CORTX-29933: System health update for control node

For control node, the node health state needs to be consumed and updated
as it is received by HA. For now, the node health state is "unknown". After,
first time, the state is marked as "starting". And then hare updates HA whether
that POD is ready or not and then HA marks it as ONLINE in system health.
But, HARE will not be informing HA regarding control node. Hence, code update is
needed when its a control node alert, consume and update the state as it is.

Signed-off-by: Madhura Mande <madhura.mande@seagate.com>

* CORTX-29933: System health update for control node

- Rebase and do changes required to so that functional_type will be present
  in the updated health event

Signed-off-by: Madhura Mande <madhura.mande@seagate.com>

* CORTX-29933: System health update for control node

- Rearrange one comment in the code

Signed-off-by: Madhura Mande <madhura.mande@seagate.com>