CORTX-29667 : HA configuration upload is in-efficient
HA maintains a part of Cortx cluster configuration in Consul KV store.
A set of relevant key-values are uploaded as part of configuration during
deployment. The present upload mechanism is in-efficient, as a single key
value is ulploaded per put request. Alternatively, as Consul supports batch
operations, several key-values can be uploaded in a single PUT transaction.

Solution:
- Use Consul transactional PUT operation to upload several HA key-values.

Signed-off-by: Mukhtar Inamdar <mukhtar.p.inamdar@seagate.com>