CORTX-31529: Handle FQDN based pod identifiers in HA mini provisioning
This ticket is needed to address the changes needed in HA in response to
changes in cluster deployment. k8s monitor service uses machine-id under
the label cortx.io/machine-id to monitor pods. And HA used to fetch this list of
machine-ids from Gconf and then it was used to pass it to monitor for match.
This same machine id was getting used through out HA to map incoming alerts in
system health. But, now, value under label and actual machine-id is different. Hence,
two things are required.
- Created list of ids that k8s monitor is required to match which will be available in
  Gconf with "node_id" key.(This if for now kept for backward compatibility)
- Afterwards, cortx.io/machine-id will not be present. Instead, cortx.io/pod_name label
  will be present and that value will be available in Gconf with field as "name". And to
  match alerts in HA, name to machine-id mapping is required. So, created that mapping
  dictionary
So, both this will be available to use with the help consul kv get.

Signed-off-by: Madhura Mande <madhura.mande@seagate.com>